Work on Project 13

+++ controllers/clothingItems.js
@@ -44,30 +44,37 @@ const createClothingItem = (req, res) => {
 
 const deleteClothingItem = (req, res) => {
   console.log("trying to delete item");
-  const { itemId } = req.params;
+  const { itemId, owner, _id } = req.params;
 
-  ClothingItem.findByIdAndRemove(itemId)
-    .orFail(() => {
-      const error = new Error("Item ID not found");
-      throw error;
-    })
-    .then(() => res.status(200).send({ message: "Item Deleted" }))
-    .catch((err) => {
-      console.error(err);
-      console.log(err.name);
-      if (err.name === "CastError") {
-        return res
-          .status(invalidDataError.status)
-          .send({ message: invalidDataError.message });
-      } if (err.name === "Error") {
-        return res
-          .status(dataNotFoundError.status)
-          .send({ message: dataNotFoundError.message });
-      }
-      return res.status(defaultError.status).send({
-        message: defaultError.message,
+  if (owner === _id) {
+    ClothingItem.findByIdAndRemove(itemId)
+      .orFail(() => {
+        const error = new Error("Item ID not found");
+        throw error;
+      })
+      .then(() => res.status(200).send({ message: "Item Deleted" }))
+      .catch((err) => {
+        console.error(err);
+        console.log(err.name);
+        if (err.name === "CastError") {
+          return res
+            .status(invalidDataError.status)
+            .send({ message: invalidDataError.message });
+        }
+        if (err.name === "Error") {
+          return res
+            .status(dataNotFoundError.status)
+            .send({ message: dataNotFoundError.message });
+        }
+        return res.status(defaultError.status).send({
+          message: defaultError.message,
+        });
       });
-    });
+  }
+  if (owner === !_id) {
+    console.log("You're not the owner of this item");
+    return res.status(403);
+  }
 };
 
 module.exports = { getClothingItems, createClothingItem, deleteClothingItem };
